♻️  Refacto userProfile page
